Add region constraints and change SNS topic ARN relative to local region
This changes the SNS topic ARN that the resulting IAM Role ARN is emitted
to from a fixed ARN to a relative ARN, relative to the region that the
GuardDuty member role is being deployed in.

Move parameters into mappings as we don't want users to be able to change
these values.

Remove GuardDutyMemberAccountId emitted attribute as this is already
present in the record in DynamoDB
Add region constraints and change SNS topic ARN relative to local region

This changes the SNS topic ARN that the resulting IAM Role ARN is emitted
to from a fixed ARN to a relative ARN, relative to the region that the
GuardDuty member role is being deployed in.

Move parameters into mappings as we don't want users to be able to change
these values.

Remove GuardDutyMemberAccountId emitted attribute as this is already
present in the record in DynamoDB